added rg.el-like buffer management
The following changes were made:

Added a fd-dired-save-search-as-name command.  This was inspired by rg.el.  It
lets the user to change the buffer name to have the ability to list fd-dired
buffers in ibuffer.

Added fd-dired-list-searches command.  This was inspired by rg.el too.  It will
list all the previously saved searches in ibuffer buffer.

Added a custom variable fd-dired-generate-random-buffer.  This variable
controlls whether the new fd-dired buffers will have random and hidden names.  I
really prefer the find-dired approach to this, because I frequently want to go
back to previous searches.  The random generated names with space at the
beginning, make this very difficult, because emacs treats names with spaces at
the beginning as internal system buffers.  All the usual commands (for listing,
filtering, viewing and deletion) will ignore these buffers.

Added a new local binding to fd-dired buffer on M-S to change the name of a
buffer.

Added ibuffer and ibuf-ext require statements